backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
gunicorn==22.0.0
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
firebase-admin==6.5.0
google-cloud-firestore==2.16.0
google-cloud-storage==2.18.2
orjson==3.10.7
